requirements.txt
aws-cdk-lib==2.130.0
constructs==10.3.0
boto3==1.34.0
redshift-connector==2.1.3
pandas==2.1.4
pyarrow==14.0.2
python-dotenv==1.0.0
